корректировка sendData, getData
